requirements.txt
absl-py==0.8.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
appnope==0.1.0
appscript==1.0.1
asn1crypto==1.0.1
astor==0.8.0
astroid==2.3.1
astropy==3.2.2
atomicwrites==1.3.0
attrs==19.2.0
Automat==0.8.0
Babel==2.7.0
backcall==0.1.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.0
bitarray==1.0.1
bkcharts==0.2
bleach==3.1.0
bokeh==1.3.4
boto==2.49.0
boto3==1.11.9
botocore==1.14.9
Bottleneck==1.2.1
cachetools==4.0.0
category-encoders==2.1.0
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
conda==4.8.2
conda-build==3.18.8
conda-package-handling==1.6.0
conda-verify==3.4.2
configparser==5.0.0
ConfigSpace==0.4.12
confuse==1.0.0
constantly==15.1.0
contextlib2==0.6.0
coverage==5.0.4
cryptography==2.7
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.13
cytoolz==0.10.0
dask==2.6.0
deap==1.3.1
decorator==4.4.0
defusedxml==0.6.0
dill==0.3.1.1
distributed==2.6.0
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
findspark==1.3.0
Flask==1.1.1
fsspec==0.5.2
future==0.17.1
gast==0.2.2
gensim==3.8.1
genson==1.2.2
geographiclib==1.50
geopy==1.20.0
gevent==1.4.0
glob2==0.7
gmpy2==2.0.8
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.51.0
greenlet==0.4.15
grpcio==1.27.2
h5py==2.9.0
HeapDict==1.0.1
holidays==0.9.11
html5lib==1.0.1
htmlmin==0.1.12
hyperlink==19.0.0
idna==2.8
imageio==2.6.0
imagesize==1.1.0
importlib-metadata==0.23
incremental==17.5.0
ipdb==0.13.4
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
jupytext==1.4.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==18.0.0
kiwisolver==1.1.0
kneed==0.7.0
lazy-object-proxy==1.4.2
liac-arff==2.4.0
libarchive-c==2.8
lief==0.9.0
llvmlite==0.29.0
locket==0.2.0
lockfile==0.12.2
logzero==1.5.0
lxml==4.4.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
numba==0.45.1
numexpr==2.7.0
numpy==1.17.3
numpydoc==0.9.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.46
openpyxl==3.0.0
opt-einsum==3.2.0
packaging==19.2
pandas==1.1.3
pandas-profiling==2.3.0
pandocfilters==1.4.2
parsel==1.5.2
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.5
patsy==0.5.1
pep8==1.7.1
pexpect==4.7.0
phik==0.9.8
pickleshare==0.7.5
Pillow==6.2.0
pkginfo==1.5.0.1
plotly==4.3.0
pluggy==0.13.0
ply==3.11
pprintpp==0.4.0
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==2.0.10
Protego==0.1.16
protobuf==3.9.2
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
py-msci-esg==0.0.8
py4j==0.10.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycountry==19.8.18
pycountry-convert==0.7.2
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
PyDispatcher==2.0.5
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.4.2
pygpu==0.7.6
PyHamcrest==2.0.0
pylint==2.4.2
pynisher==0.5.0
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyrsistent==0.15.4
PySocks==1.7.1
pyspark==2.4.5
pytest==5.2.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-cov==2.8.1
pytest-doctestplus==0.4.0
pytest-mock==3.0.0
pytest-openfiles==0.4.0
pytest-pylint==0.14.1
pytest-remotedata==0.3.2
pytest-runner==5.1
python-dateutil==2.8.0
python-forecastio==1.4.0
pytz==2019.3
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
queuelib==1.5.0
ramp-workflow==0+unknown
repoze.lru==0.7
requests==2.22.0
requests-oauthlib==1.3.0
responses==0.10.6
retrying==1.3.3
rope==0.14.0
rsa==4.0
ruamel-yaml==0.15.46
s3transfer==0.3.2
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.4.1
Scrapy==1.8.0
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
simplejson==3.17.2
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
smart-open==1.9.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
Sphinx==2.2.0
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.9
statsmodels==0.10.1
stop-words==2018.7.23
sympy==1.4
tables==3.5.2
tabpy==2.2.0
tblib==1.4.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-datasets==2.1.0
tensorflow-estimator==1.14.0
tensorflow-metadata==0.21.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
textblob==0.15.3
Theano==1.0.4
toolz==0.10.0
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
Twisted==19.10.0
typed-ast==1.4.0
typing==3.7.4.1
ujson==4.0.1
unicodecsv==0.14.1
urllib3==1.24.2
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wordcloud==1.6.0
wrapt==1.11.2
wurlitzer==1.0.3
xlrd==1.2.0
XlsxWriter==1.2.1
xlwings==0.15.10
xlwt==1.3.0
zict==1.0.0
zipp==0.6.0
zope.interface==4.7.1
